ny_tx_project/requirements.txt
pandas
pyarrow
python-dotenv
sqlalchemy
psycopg2